requirements.txt
openai-whisper
ffmpeg
gTTS
langgraph
python-dotenv
request